Final assembly
This folder will contain the full implementation of the web-application using the MVC pattern.

The login works, the db is connected, loging.vue was changed to accept and send the data. If there is a mismatch, an alert will be shown and the fields will reset.
The tree structure will change in future steps.