Fixes and features
SQL: New triggers, functions and tables. They allow to upload and download attachments and comments from tickets.
Frontend: Many small changes in extended_item, it decodes and encodes the data when posting/getting. Big, challenge was to display it for the user. currently it is not refreshing the chat on the fly, rather after a refresh.
Backend: Tickets.js were reworked for getting and sending the comments and relevant data for later use.